LPD-56967 test: move all tests related to root model to one test
