+++ modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/service/test/ObjectRelationshipLocalServiceTest.java
@@ -7,7 +7,6 @@ package com.liferay.object.service.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationRegistryUtil;
 import com.liferay.info.collection.provider.RelatedInfoItemCollectionProvider;
 import com.liferay.object.constants.ObjectActionExecutorConstants;
 import com.liferay.object.constants.ObjectActionKeys;
@@ -41,7 +40,6 @@ import com.liferay.object.model.ObjectLayoutColumn;
 import com.liferay.object.model.ObjectLayoutRow;
 import com.liferay.object.model.ObjectLayoutTab;
 import com.liferay.object.model.ObjectRelationship;
-import com.liferay.object.related.models.test.util.ObjectEntryTestUtil;
 import com.liferay.object.relationship.util.ObjectRelationshipUtil;
 import com.liferay.object.service.ObjectActionLocalService;
 import com.liferay.object.service.ObjectDefinitionLocalService;
@@ -65,7 +63,6 @@ import com.liferay.object.tree.ObjectDefinitionTreeFactory;
 import com.liferay.object.tree.Tree;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-import com.liferay.petra.function.UnsafeBiConsumer;
 import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
@@ -592,896 +589,6 @@ public class ObjectRelationshipLocalServiceTest {
 			_systemObjectDefinition2, true);
 
 		_testSystemObjectRelationshipOneToMany();
-	}
-
-	@Test
-	public void testBindDraftObjectDefinitionAndPublishedObjectDefinition()
-		throws Exception {
-
-		// Bind a draft object definition as a child node in a published object
-		// definition tree
-
-		ObjectDefinition objectDefinitionAA =
-			_addAndPublishCustomObjectDefinition("AA");
-		ObjectDefinition objectDefinitionAAA =
-			_addAndPublishCustomObjectDefinition("AAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAA.getObjectDefinitionId(),
-			objectDefinitionAAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAAA,
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAAA"),
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(2, "C_AA");
-		_asserScreenNavigationCategories(1, "C_AAA");
-		_asserScreenNavigationCategories(0, "C_AAAA");
-
-		// Bind a draft object definition as a parent node in a published
-		// object definition tree
-
-		_testBindObjectDefinitions(
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A"),
-			objectDefinitionAA,
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(0, "C_A");
-		_asserScreenNavigationCategories(2, "C_AA");
-		_asserScreenNavigationCategories(1, "C_AAA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind a draft object definition to a published object definition
-
-		_testBindObjectDefinitions(
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A"),
-			_addAndPublishCustomObjectDefinition("AA"),
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(0, "C_A");
-		_asserScreenNavigationCategories(1, "C_AA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService, new String[] {"C_A", "C_AA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind a draft object definition to a published object definition
-		// with object entries
-
-		objectDefinitionAA = _addAndPublishCustomObjectDefinition("AA");
-
-		_addObjectEntry(objectDefinitionAA, Collections.emptyMap());
-
-		ObjectDefinition objectDefinitionA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A");
-
-		_testBindObjectDefinitions(
-			objectDefinitionA, objectDefinitionAA,
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(0, "C_A");
-		_asserScreenNavigationCategories(1, "C_AA");
-
-		long objectDefinitionId = objectDefinitionA.getObjectDefinitionId();
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			StringBundler.concat(
-				"There must be no unrelated object entries when both object ",
-				"definitions are published so that the object relationship ",
-				"can be an edge to a root context"),
-			() -> _objectDefinitionLocalService.publishCustomObjectDefinition(
-				TestPropsValues.getUserId(), objectDefinitionId));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService, new String[] {"C_A", "C_AA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind a draft object definition tree to a published object definition
-		// tree
-
-		objectDefinitionA = ObjectDefinitionTestUtil.addCustomObjectDefinition(
-			"A");
-		objectDefinitionAA = ObjectDefinitionTestUtil.addCustomObjectDefinition(
-			"AA");
-
-		_bindObjectDefinitions(
-			objectDefinitionA.getObjectDefinitionId(),
-			objectDefinitionAA.getObjectDefinitionId());
-
-		objectDefinitionAAA = _addAndPublishCustomObjectDefinition("AAA");
-		ObjectDefinition objectDefinitionAAAA =
-			_addAndPublishCustomObjectDefinition("AAAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAAA.getObjectDefinitionId(),
-			objectDefinitionAAAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAA, objectDefinitionAAA,
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(0, "C_A");
-		_asserScreenNavigationCategories(0, "C_AA");
-		_asserScreenNavigationCategories(2, "C_AAA");
-		_asserScreenNavigationCategories(1, "C_AAAA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind a published object definition to a draft object definition tree
-
-		objectDefinitionA = ObjectDefinitionTestUtil.addCustomObjectDefinition(
-			"A");
-		objectDefinitionAA = ObjectDefinitionTestUtil.addCustomObjectDefinition(
-			"AA");
-
-		_bindObjectDefinitions(
-			objectDefinitionA.getObjectDefinitionId(),
-			objectDefinitionAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAA, _addAndPublishCustomObjectDefinition("AAA"),
-			(objectDefinition1, objectDefinition2) -> {
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService);
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition2.getObjectDefinitionId()),
-					_objectDefinitionLocalService);
-			});
-
-		_asserScreenNavigationCategories(0, "C_A");
-		_asserScreenNavigationCategories(0, "C_AA");
-		_asserScreenNavigationCategories(1, "C_AAA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA"}, _objectEntryLocalService,
-			_objectRelationshipLocalService);
-	}
-
-	@Test
-	public void testBindDraftObjectDefinitions() throws Exception {
-
-		// Bind a draft object definition as a child node in a draft object
-		// definition tree
-
-		ObjectDefinition objectDefinitionAA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AA");
-		ObjectDefinition objectDefinitionAAA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAA.getObjectDefinitionId(),
-			objectDefinitionAAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAAA,
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAAA"),
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		// Bind a draft object definition as a parent node in a draft object
-		// definition tree
-
-		_testBindObjectDefinitions(
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A"),
-			objectDefinitionAA,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind two draft object definition trees
-
-		ObjectDefinition objectDefinitionA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A");
-		objectDefinitionAA = ObjectDefinitionTestUtil.addCustomObjectDefinition(
-			"AA");
-
-		_bindObjectDefinitions(
-			objectDefinitionA.getObjectDefinitionId(),
-			objectDefinitionAA.getObjectDefinitionId());
-
-		objectDefinitionAAA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAA");
-		ObjectDefinition objectDefinitionAAAA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAAA.getObjectDefinitionId(),
-			objectDefinitionAAAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAA, objectDefinitionAAA,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind two draft object definitions
-
-		_testBindObjectDefinitions(
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("A"),
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AA"),
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService, new String[] {"C_A", "C_AA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-	}
-
-	@Test
-	public void testBindObjectDefinitionsWithGreaterThanTreeMaxHeight()
-		throws Exception {
-
-		// Bind an object definition to a tree that has reached the maximum
-		// height
-
-		TreeTestUtil.createObjectDefinitionTree(
-			_objectDefinitionLocalService, _objectRelationshipLocalService,
-			false,
-			LinkedHashMapBuilder.put(
-				"A", new String[] {"AA"}
-			).put(
-				"AA", new String[] {"AAA"}
-			).put(
-				"AAA", new String[] {"AAAA"}
-			).put(
-				"AAAA", new String[] {"AAAAA"}
-			).put(
-				"AAAAA", new String[0]
-			).build());
-
-		ObjectDefinition objectDefinitionAAAAA =
-			_objectDefinitionLocalService.fetchObjectDefinition(
-				TestPropsValues.getCompanyId(), "C_AAAAA");
-		ObjectDefinition objectDefinitionAAAAAA =
-			ObjectDefinitionTestUtil.addCustomObjectDefinition("AAAAAA");
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			"The object relationship cannot be an edge in the root context " +
-				"because it would exceed the tree's maximum height",
-			() -> _bindObjectDefinitions(
-				objectDefinitionAAAAA.getObjectDefinitionId(),
-				objectDefinitionAAAAAA.getObjectDefinitionId()));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA", "C_AAAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		_objectDefinitionLocalService.deleteObjectDefinition(
-			objectDefinitionAAAAAA);
-
-		// Bind two object definition trees into one so that the height
-		// of the new tree exceeds the maximum height
-
-		TreeTestUtil.createObjectDefinitionTree(
-			_objectDefinitionLocalService, _objectRelationshipLocalService,
-			false,
-			LinkedHashMapBuilder.put(
-				"A", new String[] {"AA"}
-			).put(
-				"AA", new String[] {"AAA"}
-			).put(
-				"AAA", new String[0]
-			).build());
-
-		TreeTestUtil.createObjectDefinitionTree(
-			_objectDefinitionLocalService, _objectRelationshipLocalService,
-			false,
-			LinkedHashMapBuilder.put(
-				"AAAA", new String[] {"AAAAA"}
-			).put(
-				"AAAAA", new String[] {"AAAAAA"}
-			).put(
-				"AAAAAA", new String[0]
-			).build());
-
-		ObjectDefinition objectDefinitionAAA =
-			_objectDefinitionLocalService.fetchObjectDefinition(
-				TestPropsValues.getCompanyId(), "C_AAA");
-		ObjectDefinition objectDefinitionAAAA =
-			_objectDefinitionLocalService.fetchObjectDefinition(
-				TestPropsValues.getCompanyId(), "C_AAAA");
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			"The object relationship cannot be an edge in the root context " +
-				"because it would exceed the tree's maximum height",
-			() -> _bindObjectDefinitions(
-				objectDefinitionAAA.getObjectDefinitionId(),
-				objectDefinitionAAAA.getObjectDefinitionId()));
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {
-				"C_A", "C_AA", "C_AAA", "C_AAAA", "C_AAAAA", "C_AAAAAA"
-			},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-	}
-
-	@Test
-	public void testBindPublishedObjectDefinitions() throws Exception {
-
-		// Bind a published object definition as a child node in a published
-		// object definition tree
-
-		ObjectDefinition objectDefinitionAA =
-			_addAndPublishCustomObjectDefinition("AA");
-		ObjectDefinition objectDefinitionAAA =
-			_addAndPublishCustomObjectDefinition("AAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAA.getObjectDefinitionId(),
-			objectDefinitionAAA.getObjectDefinitionId());
-
-		ObjectDefinition objectDefinitionAAAA =
-			_addAndPublishCustomObjectDefinition("AAAA");
-
-		_testBindObjectDefinitions(
-			objectDefinitionAAA, objectDefinitionAAAA,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		_asserScreenNavigationCategories(2, "C_AA");
-		_asserScreenNavigationCategories(2, "C_AAA");
-		_asserScreenNavigationCategories(1, "C_AAAA");
-
-		// Bind a published object definition as a parent node in a published
-		// object definition tree
-
-		_testBindObjectDefinitions(
-			_addAndPublishCustomObjectDefinition("A"), objectDefinitionAA,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		_asserScreenNavigationCategories(2, "C_A");
-		_asserScreenNavigationCategories(2, "C_AA");
-		_asserScreenNavigationCategories(2, "C_AAA");
-		_asserScreenNavigationCategories(1, "C_AAAA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind two published object definition trees
-
-		ObjectDefinition objectDefinitionA =
-			_addAndPublishCustomObjectDefinition("A");
-		objectDefinitionAA = _addAndPublishCustomObjectDefinition("AA");
-
-		_bindObjectDefinitions(
-			objectDefinitionA.getObjectDefinitionId(),
-			objectDefinitionAA.getObjectDefinitionId());
-
-		objectDefinitionAAA = _addAndPublishCustomObjectDefinition("AAA");
-		objectDefinitionAAAA = _addAndPublishCustomObjectDefinition("AAAA");
-
-		_bindObjectDefinitions(
-			objectDefinitionAAA.getObjectDefinitionId(),
-			objectDefinitionAAAA.getObjectDefinitionId());
-
-		_testBindObjectDefinitions(
-			objectDefinitionAA, objectDefinitionAAA,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"A", new String[] {"AA"}
-					).put(
-						"AA", new String[] {"AAA"}
-					).put(
-						"AAA", new String[] {"AAAA"}
-					).put(
-						"AAAA", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getRootObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		_asserScreenNavigationCategories(2, "C_A");
-		_asserScreenNavigationCategories(2, "C_AA");
-		_asserScreenNavigationCategories(2, "C_AAA");
-		_asserScreenNavigationCategories(1, "C_AAAA");
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_A", "C_AA", "C_AAA", "C_AAAA"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-
-		// Bind two published object definitions
-
-		ObjectDefinition objectDefinitionB =
-			_addAndPublishCustomObjectDefinition("B");
-		ObjectDefinition objectDefinitionBB =
-			_addAndPublishCustomObjectDefinition("BB");
-
-		_testBindObjectDefinitions(
-			objectDefinitionB, objectDefinitionBB,
-			(objectDefinition1, objectDefinition2) ->
-				TreeTestUtil.assertObjectDefinitionTree(
-					LinkedHashMapBuilder.put(
-						"B", new String[] {"BB"}
-					).put(
-						"BB", new String[0]
-					).build(),
-					_objectDefinitionTreeFactory.create(
-						objectDefinition1.getObjectDefinitionId()),
-					_objectDefinitionLocalService));
-
-		_asserScreenNavigationCategories(2, "C_B");
-		_asserScreenNavigationCategories(1, "C_BB");
-
-		// Object definitions must have the same scope to enable inheritance
-
-		ObjectDefinition siteObjectDefinition =
-			ObjectDefinitionTestUtil.publishObjectDefinition(
-				Collections.singletonList(
-					new TextObjectFieldBuilder(
-					).labelMap(
-						LocalizedMapUtil.getLocalizedMap(
-							RandomTestUtil.randomString())
-					).name(
-						"a" + RandomTestUtil.randomString()
-					).build()),
-				ObjectDefinitionConstants.SCOPE_SITE);
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			String.format(
-				"The scope of \"%s\" is not the same as \"%s\". To enable " +
-					"inheritance, the object definitions must have the same " +
-						"scope",
-				objectDefinitionB.getShortName(),
-				siteObjectDefinition.getShortName()),
-			() -> _bindObjectDefinitions(
-				objectDefinitionB.getObjectDefinitionId(),
-				siteObjectDefinition.getObjectDefinitionId()));
-
-		// Unable to bind the object definitions because the object relationship
-		// must not create a circular reference in a root context
-
-		ObjectDefinition objectDefinitionBBB =
-			_addAndPublishCustomObjectDefinition("BBB");
-
-		_bindObjectDefinitions(
-			objectDefinitionBB.getObjectDefinitionId(),
-			objectDefinitionBBB.getObjectDefinitionId());
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			"The object relationship must not create a circular reference in " +
-				"a root context",
-			() -> _bindObjectDefinitions(
-				objectDefinitionBBB.getObjectDefinitionId(),
-				objectDefinitionB.getObjectDefinitionId()));
-
-		ObjectDefinition objectDefinitionC =
-			_addAndPublishCustomObjectDefinition("C");
-		ObjectDefinition objectDefinitionCC =
-			_addAndPublishCustomObjectDefinition("CC");
-
-		_bindObjectDefinitions(
-			objectDefinitionC.getObjectDefinitionId(),
-			objectDefinitionCC.getObjectDefinitionId());
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {"C_B", "C_BB", "C_BBB"}, _objectEntryLocalService,
-			_objectRelationshipLocalService);
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService, new String[] {"C_C", "C_CC"},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-	}
-
-	@Test
-	public void testBindPublishedObjectDefinitionsWithObjectEntries()
-		throws Exception {
-
-		ObjectDefinition objectDefinitionAA =
-			_addAndPublishCustomObjectDefinition("AA");
-		ObjectDefinition objectDefinitionAAA =
-			_addAndPublishCustomObjectDefinition("AAA");
-
-		ObjectRelationship objectRelationship1 =
-			_objectRelationshipLocalService.addObjectRelationship(
-				StringUtil.randomId(), TestPropsValues.getUserId(),
-				objectDefinitionAA.getObjectDefinitionId(),
-				objectDefinitionAAA.getObjectDefinitionId(), 0,
-				ObjectRelationshipConstants.DELETION_TYPE_PREVENT, false,
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				StringUtil.randomId(), false,
-				ObjectRelationshipConstants.TYPE_ONE_TO_MANY, null);
-
-		ObjectEntry objectDefinitionAAObjectEntry1 =
-			ObjectEntryTestUtil.addObjectEntry(
-				0, objectDefinitionAA.getObjectDefinitionId(),
-				Collections.emptyMap());
-		ObjectEntry objectDefinitionAAAObjectEntry1 =
-			ObjectEntryTestUtil.addObjectEntry(
-				0, objectDefinitionAAA.getObjectDefinitionId(),
-				Collections.emptyMap());
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			StringBundler.concat(
-				"There must be no unrelated object entries when both object ",
-				"definitions are published so that the object relationship ",
-				"can be an edge to a root context"),
-			() -> _bindObjectDefinitions(objectRelationship1));
-
-		ObjectField objectField1 = _objectFieldLocalService.getObjectField(
-			objectRelationship1.getObjectFieldId2());
-
-		objectDefinitionAAAObjectEntry1 =
-			_objectEntryLocalService.updateObjectEntry(
-				objectDefinitionAAAObjectEntry1.getUserId(),
-				objectDefinitionAAAObjectEntry1.getObjectEntryId(),
-				Collections.singletonMap(
-					objectField1.getName(),
-					objectDefinitionAAObjectEntry1.getObjectEntryId()),
-				ServiceContextTestUtil.getServiceContext());
-
-		Assert.assertEquals(
-			0, objectDefinitionAAObjectEntry1.getRootObjectEntryId());
-		Assert.assertEquals(
-			0, objectDefinitionAAAObjectEntry1.getRootObjectEntryId());
-
-		_bindObjectDefinitions(objectRelationship1);
-
-		objectDefinitionAAObjectEntry1 =
-			_objectEntryLocalService.getObjectEntry(
-				objectDefinitionAAObjectEntry1.getObjectEntryId());
-		objectDefinitionAAAObjectEntry1 =
-			_objectEntryLocalService.getObjectEntry(
-				objectDefinitionAAAObjectEntry1.getObjectEntryId());
-
-		long expectedRootObjectEntryId =
-			objectDefinitionAAObjectEntry1.getRootObjectEntryId();
-
-		Assert.assertEquals(
-			expectedRootObjectEntryId,
-			objectDefinitionAAObjectEntry1.getRootObjectEntryId());
-		Assert.assertEquals(
-			expectedRootObjectEntryId,
-			objectDefinitionAAAObjectEntry1.getRootObjectEntryId());
-
-		ObjectDefinition objectDefinitionA =
-			_addAndPublishCustomObjectDefinition("A");
-
-		ObjectRelationship objectRelationship2 =
-			_objectRelationshipLocalService.addObjectRelationship(
-				StringUtil.randomId(), TestPropsValues.getUserId(),
-				objectDefinitionA.getObjectDefinitionId(),
-				objectDefinitionAA.getObjectDefinitionId(), 0,
-				ObjectRelationshipConstants.DELETION_TYPE_PREVENT, false,
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				StringUtil.randomId(), false,
-				ObjectRelationshipConstants.TYPE_ONE_TO_MANY, null);
-
-		ObjectField objectField2 = _objectFieldLocalService.getObjectField(
-			objectRelationship2.getObjectFieldId2());
-
-		ObjectEntry objectDefinitionAObjectEntry1 =
-			ObjectEntryTestUtil.addObjectEntry(
-				0, objectDefinitionA.getObjectDefinitionId(),
-				Collections.emptyMap());
-
-		_objectEntryLocalService.updateObjectEntry(
-			objectDefinitionAAObjectEntry1.getUserId(),
-			objectDefinitionAAObjectEntry1.getObjectEntryId(),
-			Collections.singletonMap(
-				objectField2.getName(),
-				objectDefinitionAObjectEntry1.getObjectEntryId()),
-			ServiceContextTestUtil.getServiceContext());
-
-		Assert.assertEquals(
-			0, objectDefinitionAObjectEntry1.getRootObjectEntryId());
-
-		_bindObjectDefinitions(objectRelationship2);
-
-		objectDefinitionAObjectEntry1 = _objectEntryLocalService.getObjectEntry(
-			objectDefinitionAObjectEntry1.getObjectEntryId());
-		objectDefinitionAAObjectEntry1 =
-			_objectEntryLocalService.getObjectEntry(
-				objectDefinitionAAObjectEntry1.getObjectEntryId());
-		objectDefinitionAAAObjectEntry1 =
-			_objectEntryLocalService.getObjectEntry(
-				objectDefinitionAAAObjectEntry1.getObjectEntryId());
-
-		expectedRootObjectEntryId =
-			objectDefinitionAObjectEntry1.getRootObjectEntryId();
-
-		Assert.assertEquals(
-			expectedRootObjectEntryId,
-			objectDefinitionAObjectEntry1.getRootObjectEntryId());
-		Assert.assertEquals(
-			expectedRootObjectEntryId,
-			objectDefinitionAAObjectEntry1.getRootObjectEntryId());
-		Assert.assertEquals(
-			expectedRootObjectEntryId,
-			objectDefinitionAAAObjectEntry1.getRootObjectEntryId());
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {
-				objectDefinitionA.getName(), objectDefinitionAA.getName(),
-				objectDefinitionAAA.getName()
-			},
-			_objectEntryLocalService, _objectRelationshipLocalService);
-	}
-
-	@Test
-	public void testBindPublishedObjectDefinitionsWithObjectEntriesAndWorkflowInstance()
-		throws Exception {
-
-		ObjectDefinition objectDefinitionA =
-			_addAndPublishCustomObjectDefinition();
-		ObjectDefinition objectDefinitionAA =
-			_addAndPublishCustomObjectDefinition();
-
-		ObjectRelationship objectRelationship =
-			_objectRelationshipLocalService.addObjectRelationship(
-				StringUtil.randomId(), TestPropsValues.getUserId(),
-				objectDefinitionA.getObjectDefinitionId(),
-				objectDefinitionAA.getObjectDefinitionId(), 0,
-				ObjectRelationshipConstants.DELETION_TYPE_PREVENT, false,
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				StringUtil.randomId(), false,
-				ObjectRelationshipConstants.TYPE_ONE_TO_MANY, null);
-
-		_workflowDefinitionLinkLocalService.updateWorkflowDefinitionLink(
-			TestPropsValues.getUserId(), TestPropsValues.getCompanyId(), 0,
-			objectDefinitionA.getClassName(), 0, 0, "Single Approver", 1);
-
-		ObjectEntry objectEntryA1 = ObjectEntryTestUtil.addObjectEntry(
-			0, objectDefinitionA.getObjectDefinitionId(),
-			Collections.emptyMap());
-		ObjectEntry objectEntryA2 = ObjectEntryTestUtil.addObjectEntry(
-			0, objectDefinitionA.getObjectDefinitionId(),
-			Collections.emptyMap());
-
-		_workflowDefinitionLinkLocalService.updateWorkflowDefinitionLink(
-			TestPropsValues.getUserId(), TestPropsValues.getCompanyId(), 0,
-			objectDefinitionAA.getClassName(), 0, 0, "Single Approver", 1);
-
-		ObjectField objectField = _objectFieldLocalService.getObjectField(
-			objectRelationship.getObjectFieldId2());
-
-		ObjectEntry objectEntryAA = ObjectEntryTestUtil.addObjectEntry(
-			0, objectDefinitionAA.getObjectDefinitionId(),
-			Collections.singletonMap(
-				objectField.getName(), objectEntryA1.getObjectEntryId()));
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			String.format(
-				"These ongoing workflow instances must be completed to " +
-					"enable inheritance: \"%s\" (\"%s\" object entries) and " +
-						"\"%s\" (\"%s\" object entries)",
-				objectDefinitionA.getLabel(LocaleUtil.US), 2,
-				objectDefinitionAA.getLabel(LocaleUtil.US), 1),
-			() -> _bindObjectDefinitions(objectRelationship));
-
-		_completeWorkflowTask(
-			objectDefinitionA.getClassName(), objectEntryA1.getObjectEntryId());
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			String.format(
-				"These ongoing workflow instances must be completed to " +
-					"enable inheritance: \"%s\" (\"%s\" object entries) and " +
-						"\"%s\" (\"%s\" object entries)",
-				objectDefinitionA.getLabel(LocaleUtil.US), 1,
-				objectDefinitionAA.getLabel(LocaleUtil.US), 1),
-			() -> _bindObjectDefinitions(objectRelationship));
-
-		_completeWorkflowTask(
-			objectDefinitionA.getClassName(), objectEntryA2.getObjectEntryId());
-
-		AssertUtils.assertFailure(
-			ObjectRelationshipEdgeException.class,
-			String.format(
-				"These ongoing workflow instances must be completed to " +
-					"enable inheritance: \"%s\" (\"%s\" object entries)",
-				objectDefinitionAA.getLabel(LocaleUtil.US), 1),
-			() -> _bindObjectDefinitions(objectRelationship));
-
-		_completeWorkflowTask(
-			objectDefinitionAA.getClassName(),
-			objectEntryAA.getObjectEntryId());
-
-		TreeTestUtil.bind(
-			_objectRelationshipLocalService, List.of(objectRelationship));
-
-		objectEntryAA = _objectEntryLocalService.getObjectEntry(
-			objectEntryAA.getObjectEntryId());
-
-		Assert.assertEquals(
-			objectEntryA1.getObjectEntryId(),
-			objectEntryAA.getRootObjectEntryId());
-
-		TreeTestUtil.deleteObjectDefinitionHierarchy(
-			_objectDefinitionLocalService,
-			new String[] {
-				objectDefinitionA.getName(), objectDefinitionAA.getName()
-			},
-			_objectEntryLocalService, _objectRelationshipLocalService);
 	}
 
 	@Test
@@ -2332,22 +1439,17 @@ public class ObjectRelationshipLocalServiceTest {
 	private ObjectDefinition _addAndPublishModifiableSystemObjectDefinition()
 		throws Exception {
 
-		ObjectDefinition modifiableSystemObjectDefinition =
-			ObjectDefinitionTestUtil.addModifiableSystemObjectDefinition(
-				TestPropsValues.getUserId(), null, false,
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				"Test", null, null,
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				ObjectDefinitionConstants.SCOPE_SITE, null, 1,
-				Arrays.asList(
-					ObjectFieldUtil.createObjectField(
-						ObjectFieldConstants.BUSINESS_TYPE_TEXT,
-						ObjectFieldConstants.DB_TYPE_STRING,
-						RandomTestUtil.randomString(), StringUtil.randomId())));
-
-		return _objectDefinitionLocalService.publishSystemObjectDefinition(
-			TestPropsValues.getUserId(),
-			modifiableSystemObjectDefinition.getObjectDefinitionId());
+		return ObjectDefinitionTestUtil.addModifiableSystemObjectDefinition(
+			TestPropsValues.getUserId(), null, false,
+			LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
+			"Test", null, null,
+			LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
+			ObjectDefinitionConstants.SCOPE_SITE, null, 1,
+			Arrays.asList(
+				ObjectFieldUtil.createObjectField(
+					ObjectFieldConstants.BUSINESS_TYPE_TEXT,
+					ObjectFieldConstants.DB_TYPE_STRING,
+					RandomTestUtil.randomString(), StringUtil.randomId())));
 	}
 
 	private ObjectEntry _addObjectEntry(
@@ -2458,24 +1560,6 @@ public class ObjectRelationshipLocalServiceTest {
 			ObjectRelationshipConstants.TYPE_ONE_TO_MANY, null);
 	}
 
-	private void _asserScreenNavigationCategories(
-			int expectedSize, String objectDefinitionName)
-		throws Exception {
-
-		ObjectDefinition objectDefinition =
-			_objectDefinitionLocalService.getObjectDefinition(
-				TestPropsValues.getCompanyId(), objectDefinitionName);
-
-		List<ScreenNavigationCategory> screenNavigationCategories =
-			ScreenNavigationRegistryUtil.getScreenNavigationCategories(
-				objectDefinition.getClassName(), TestPropsValues.getUser(),
-				null);
-
-		Assert.assertEquals(
-			screenNavigationCategories.toString(), expectedSize,
-			screenNavigationCategories.size());
-	}
-
 	private void _assertHasResourcePermission(
 			boolean expectedHasResourcePermission, ObjectEntry objectEntry,
 			Role role)
@@ -2542,19 +1626,6 @@ public class ObjectRelationshipLocalServiceTest {
 
 		Assert.assertEquals(
 			expectedRootObjectEntryId, objectEntry.getRootObjectEntryId());
-	}
-
-	private ObjectRelationship _bindObjectDefinitions(
-			long objectDefinitionId1, long objectDefinitionId2)
-		throws Exception {
-
-		return _objectRelationshipLocalService.addObjectRelationship(
-			StringUtil.randomId(), TestPropsValues.getUserId(),
-			objectDefinitionId1, objectDefinitionId2, 0,
-			ObjectRelationshipConstants.DELETION_TYPE_CASCADE, true,
-			LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-			StringUtil.randomId(), false,
-			ObjectRelationshipConstants.TYPE_ONE_TO_MANY, null);
 	}
 
 	private ObjectRelationship _bindObjectDefinitions(
@@ -2788,29 +1859,6 @@ public class ObjectRelationshipLocalServiceTest {
 
 		_objectRelationshipLocalService.deleteObjectRelationship(
 			objectRelationship);
-	}
-
-	private void _testBindObjectDefinitions(
-			ObjectDefinition objectDefinition1,
-			ObjectDefinition objectDefinition2,
-			UnsafeBiConsumer<ObjectDefinition, ObjectDefinition, Exception>
-				biConsumer)
-		throws Exception {
-
-		ObjectRelationship objectRelationship = _bindObjectDefinitions(
-			objectDefinition1.getObjectDefinitionId(),
-			objectDefinition2.getObjectDefinitionId());
-
-		Assert.assertTrue(objectRelationship.isEdge());
-		Assert.assertEquals(
-			ObjectRelationshipConstants.DELETION_TYPE_CASCADE,
-			objectRelationship.getDeletionType());
-
-		biConsumer.accept(
-			_objectDefinitionLocalService.getObjectDefinition(
-				objectDefinition1.getObjectDefinitionId()),
-			_objectDefinitionLocalService.getObjectDefinition(
-				objectDefinition2.getObjectDefinitionId()));
 	}
 
 	private void _testCreateManyToManyObjectRelationshipTable(
